Added bucket list window sample period config setting

+++ src/bucket/Bucket.cpp
@@ -838,7 +838,7 @@ mergeCasesWithEqualKeys(MergeCounters& mc, BucketInputIterator& oi,
 
 bool
 Bucket::scanForEviction(AbstractLedgerTxn& ltx, EvictionIterator& iter,
-                        uint64_t& bytesToScan,
+                        uint32_t& bytesToScan,
                         uint32_t& remainingEntriesToEvict, uint32_t ledgerSeq,
                         medida::Counter& entriesEvictedCounter,
                         medida::Counter& bytesScannedForEvictionCounter,
